+++ db/migrate/20150722062644_add_manifestation_id_to_toc.rb
@@ -1,6 +1,6 @@
 class AddManifestationIdToToc < ActiveRecord::Migration[7.1]
   def change
-    add_column :tocs, :manifestation_id, :integer
+    add_column :tocs, :manifestation_id, :bigint
     add_index :tocs, :manifestation_id
   end
 end
